R2 Logs 100daysOfCode Challenge
Today is August 3, 2019 and I have decided to start the round 2 of the challenge. I went on doing some job searches for software or web developers and one of the job that I tried to apply for was asking on whether I can recreate a website from a pdf format file. If that means from an existing design that needs to be coded so I've decided to restart the landing page challenge of my FCC challenge and try to reverse engineer the site in codepen.