Round2 Day7 August 10, 2019
Did some troubleshooting with the issue of centering the text in the product section . Found out that the list element is adapting the float property of the nav that I gave it to float on the left. So I just applied float: none; value to solve the issue of the text not aligning to the center.